+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/InMemoryFilmStorage.java
@@ -6,7 +6,6 @@ import ru.yandex.practicum.filmorate.exception.NewFilmException;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
 
-import javax.validation.Valid;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -16,7 +15,7 @@ import java.util.Map;
 @Slf4j
 @Component
 public class InMemoryFilmStorage implements FilmStorage {
-    Map<Integer, Film> films = new HashMap<>();
+    Map<Long, Film> films = new HashMap<>();
     private int idCounter = 1;
 
     public List<Film> getAllFilms() {
@@ -51,7 +50,7 @@ public class InMemoryFilmStorage implements FilmStorage {
         return film;
     }
 
-    public void deleteFilm(int id) {
+    public void deleteFilm(long id) {
         if (films.containsKey(id)) {
             films.remove(id);
         } else log.warn("фильма с id = " + id + " не существует");
